+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [Flatten](https://fsharp.github.io/fsharp-core-docs/reference/fsharp-core-optionmodule.html#flatten) function
 - [Fold](https://fsharp.github.io/fsharp-core-docs/reference/fsharp-core-optionmodule.html#fold) function
 - [FoldBack](https://fsharp.github.io/fsharp-core-docs/reference/fsharp-core-optionmodule.html#foldBack) function
+- [ForAll](https://fsharp.github.io/fsharp-core-docs/reference/fsharp-core-optionmodule.html#forall) function
 ### Changed
 - Class structure. Moved Ext class to own file
 - Test project target framework to .Net 6